dom/xbl/test/test_bug468210.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=468210
-->
<head>
  <title>Test for Bug 468210</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <!-- Keep the stuff inside <content> without spaces, so that the getAnonymousNodes works right -->
  <bindings xmlns="http://www.mozilla.org/xbl">
    <binding id="foo">
      <content><span xmlns='http://www.w3.org/1999/xhtml'><children xmlns="http://www.mozilla.org/xbl"/></span></content>
    </binding>
  </bindings>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=468210">Mozilla Bug 468210</a>
<p id="display">
  <div id="d" style="-moz-binding: url(#foo);"></div>
  <a name="foo"></a>
</p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">
<![CDATA[

/** Test for Bug 468210 **/
SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  var div = $("d");
  var n = document.anchors.length;
  is(n, 1, "Unexpected number of anchors");
  var anon = SpecialPowers.wrap(document).getAnonymousNodes(div)[0];
  ok(SpecialPowers.call_Instanceof(anon, HTMLSpanElement), "Unexpected node");
  is(SpecialPowers.unwrap(anon.parentNode), div, "Unexpected parent");
  document.body.appendChild(div);
  is(anon.parentNode, null, "Parent should have become null");
  // An attr set to test notifications
  anon.setAttribute("h", "i");
});
addLoadEvent(SimpleTest.finish);



]]>
</script>
</pre>
</body>
</html>
